Remove subcommand logging feature experiment
