+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/handlers/tekuv1/admin/PutLogLevel.java
@@ -70,7 +70,7 @@ public class PutLogLevel extends AbstractHandler implements Handler {
       final String[] logFilters = params.getLogFilter().orElseGet(() -> new String[] {""});
 
       for (final String logFilter : logFilters) {
-        LoggingConfigurator.setAllLevels(logFilter, params.getLevel(), false);
+        LoggingConfigurator.setAllLevels(logFilter, params.getLevel());
       }
 
       ctx.status(SC_NO_CONTENT);
