+++ teku/src/main/java/tech/pegasys/teku/cli/subcommand/VoluntaryExitCommand.java
@@ -26,19 +26,15 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Scanner;
 import java.util.stream.Collectors;
-import org.apache.logging.log4j.LogManager;
 import org.apache.tuweni.bytes.Bytes32;
 import org.hyperledger.besu.metrics.noop.NoOpMetricsSystem;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
 import picocli.CommandLine;
 import picocli.CommandLine.Help.Visibility;
-import picocli.CommandLine.Mixin;
 import tech.pegasys.teku.api.schema.SignedVoluntaryExit;
 import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.bls.BLSSignature;
 import tech.pegasys.teku.cli.converter.PicoCliVersionProvider;
-import tech.pegasys.teku.cli.options.DataOptions;
-import tech.pegasys.teku.cli.options.LoggingOptions;
 import tech.pegasys.teku.cli.options.ValidatorClientOptions;
 import tech.pegasys.teku.cli.options.ValidatorKeysOptions;
 import tech.pegasys.teku.config.TekuConfiguration;
@@ -47,9 +43,6 @@ import tech.pegasys.teku.infrastructure.async.AsyncRunner;
 import tech.pegasys.teku.infrastructure.async.AsyncRunnerFactory;
 import tech.pegasys.teku.infrastructure.async.MetricTrackingExecutorFactory;
 import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException;
-import tech.pegasys.teku.infrastructure.logging.LoggingConfig;
-import tech.pegasys.teku.infrastructure.logging.LoggingConfigurator;
-import tech.pegasys.teku.infrastructure.logging.LoggingDestination;
 import tech.pegasys.teku.infrastructure.logging.SubCommandLogger;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.Spec;
@@ -75,7 +68,6 @@ import tech.pegasys.teku.validator.remote.apiclient.OkHttpValidatorRestApiClient
 public class VoluntaryExitCommand implements Runnable {
   public static final SubCommandLogger SUB_COMMAND_LOG = new SubCommandLogger();
   private static final int MAX_PUBLIC_KEY_BATCH_SIZE = 50;
-  public static final String LOG_FILE_PREFIX = "teku-exit";
   private OkHttpValidatorRestApiClient apiClient;
   private tech.pegasys.teku.spec.datastructures.state.Fork fork;
   private Bytes32 genesisRoot;
@@ -89,12 +81,6 @@ public class VoluntaryExitCommand implements Runnable {
 
   @CommandLine.Mixin(name = "Validator Client")
   private ValidatorClientOptions validatorClientOptions;
-
-  @Mixin(name = "Logging")
-  private LoggingOptions loggingOptions;
-
-  @Mixin(name = "Data")
-  private DataOptions dataOptions;
 
   @CommandLine.Option(
       names = {"--epoch"},
@@ -116,7 +102,6 @@ public class VoluntaryExitCommand implements Runnable {
   public void run() {
     SUB_COMMAND_LOG.display("Loading configuration...");
     try {
-      startLogging();
       initialise();
       if (confirmationEnabled) {
         confirmExits();
@@ -131,14 +116,6 @@ public class VoluntaryExitCommand implements Runnable {
       SUB_COMMAND_LOG.error("Fatal error in VoluntaryExit. Exiting", ex);
       System.exit(1);
     }
-  }
-
-  private void startLogging() {
-    LoggingConfig loggingConfig =
-        loggingOptions.applyLoggingConfiguration(
-            dataOptions.getDataPath(), LOG_FILE_PREFIX, LoggingDestination.FILE);
-    LoggingConfigurator.update(loggingConfig, true);
-    SUB_COMMAND_LOG.setLogger(LogManager.getLogger("VoluntaryExit"));
   }
 
   private void confirmExits() {
